Remove abstract gateway, no longer used.
